[#4311] refactor results with a router switch between two views
