+++ akvo/rsr/spa/app/modules/results/enumerator.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Enumerator = () => {
+  return (
+    <div className="enumerator">
+      enumerator
+    </div>
+  )
+}
+
+export default Enumerator
